+++ src/material/paginator/testing/BUILD.bazel
@@ -17,20 +17,7 @@ ts_library(
 
 ng_test_library(
     name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = ["shared.spec.ts"],
-    ),
-    deps = [
-        ":harness_tests_lib",
-        ":testing",
-        "//src/material/paginator",
-    ],
-)
-
-ng_test_library(
-    name = "harness_tests_lib",
-    srcs = ["shared.spec.ts"],
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
         ":testing",
         "//src/cdk/testing",
